Remove the `context_stack` dependency from the `transformations` module.
PiperOrigin-RevId: 259602627
