+++ tensorflow_federated/python/core/impl/compiler_pipeline.py
@@ -25,6 +25,7 @@ from tensorflow_federated.python.core.impl import computation_building_blocks
 from tensorflow_federated.python.core.impl import computation_impl
 from tensorflow_federated.python.core.impl import context_stack_base
 from tensorflow_federated.python.core.impl import transformations
+from tensorflow_federated.python.core.impl import value_transformations
 
 
 class CompilerPipeline(object):
@@ -74,7 +75,7 @@ class CompilerPipeline(object):
     # Replace intrinsics with their bodies, for now manually in a fixed order.
     # TODO(b/113123410): Replace this with a more automated implementation that
     # does not rely on manual maintenance.
-    comp, _ = transformations.replace_all_intrinsics_with_bodies(
+    comp, _ = value_transformations.replace_all_intrinsics_with_bodies(
         comp, self._context_stack)
 
     # Replaces called lambdas with LET constructs with a single local symbol.
