+++ tensorflow_federated/python/core/impl/intrinsic_reductions.py
@@ -21,7 +21,7 @@ from __future__ import print_function
 from tensorflow_federated.python.common_libs import py_typecheck
 from tensorflow_federated.python.core.impl import computation_building_blocks
 from tensorflow_federated.python.core.impl import context_stack_impl
-from tensorflow_federated.python.core.impl import transformations
+from tensorflow_federated.python.core.impl import value_transformations
 
 
 def replace_intrinsics_with_bodies(comp):
@@ -48,6 +48,6 @@ def replace_intrinsics_with_bodies(comp):
   py_typecheck.check_type(comp,
                           computation_building_blocks.ComputationBuildingBlock)
   context_stack = context_stack_impl.context_stack
-  comp, _ = transformations.replace_all_intrinsics_with_bodies(
+  comp, _ = value_transformations.replace_all_intrinsics_with_bodies(
       comp, context_stack)
   return comp
