+++ tensorflow_federated/python/core/impl/transformations.py
@@ -30,9 +30,6 @@ from tensorflow_federated.python.core.impl import compiled_computation_transform
 from tensorflow_federated.python.core.impl import computation_building_block_utils
 from tensorflow_federated.python.core.impl import computation_building_blocks
 from tensorflow_federated.python.core.impl import computation_constructing_utils
-from tensorflow_federated.python.core.impl import context_stack_base
-from tensorflow_federated.python.core.impl import federated_computation_utils
-from tensorflow_federated.python.core.impl import intrinsic_bodies
 from tensorflow_federated.python.core.impl import intrinsic_defs
 from tensorflow_federated.python.core.impl import transformation_utils
 from tensorflow_federated.python.core.impl import tree_analysis
@@ -889,52 +886,6 @@ def replace_called_lambda_with_block(comp):
   return transformation_utils.transform_postorder(comp, _transform)
 
 
-def replace_intrinsic_with_callable(comp, uri, body, context_stack):
-  """Replaces all the intrinsics with the given `uri` with a callable.
-
-  This transform traverses `comp` postorder and replaces all the intrinsics with
-  the given `uri` with a polymorphic callable that represents the body of the
-  implementation of the intrinsic; i.e., one that given the parameter of the
-  intrinsic constructs the intended result. This will typically be a Python
-  function decorated with `@federated_computation` to make it into a polymorphic
-  callable.
-
-  Args:
-    comp: The computation building block in which to perform the replacements.
-    uri: The URI of the intrinsic to replace.
-    body: A polymorphic callable.
-    context_stack: The context stack to use.
-
-  Returns:
-    A new computation with the transformation applied or the original `comp`.
-
-  Raises:
-    TypeError: If types do not match.
-  """
-  py_typecheck.check_type(comp,
-                          computation_building_blocks.ComputationBuildingBlock)
-  py_typecheck.check_type(uri, six.string_types)
-  py_typecheck.check_type(context_stack, context_stack_base.ContextStack)
-  if not callable(body):
-    raise TypeError('The body of the intrinsic must be a callable.')
-
-  def _should_transform(comp):
-    return (isinstance(comp, computation_building_blocks.Intrinsic) and
-            comp.uri == uri and
-            isinstance(comp.type_signature, computation_types.FunctionType))
-
-  def _transform(comp):
-    if not _should_transform(comp):
-      return comp, False
-    # We need 'wrapped_body' to accept exactly one argument.
-    wrapped_body = lambda x: body(x)  # pylint: disable=unnecessary-lambda
-    transformed_comp = federated_computation_utils.zero_or_one_arg_fn_to_building_block(
-        wrapped_body, 'arg', comp.type_signature.parameter, context_stack, uri)
-    return transformed_comp, True
-
-  return transformation_utils.transform_postorder(comp, _transform)
-
-
 def replace_selection_from_tuple_with_element(comp):
   r"""Replaces any selection from a tuple with the underlying tuple element.
 
@@ -1519,35 +1470,6 @@ def unwrap_placement(comp):
       lambda_wrapping_placement_removal, ref_to_fed_arg)
 
   return called_intrinsic, True
-
-
-def replace_all_intrinsics_with_bodies(comp, context_stack):
-  """Iterates over all intrinsic bodies, inlining the intrinsics in `comp`.
-
-  Args:
-    comp: Instance of `computation_building_blocks.ComputationBuildingBlock` in
-      which we wish to replace all intrinsics with their bodies.
-    context_stack: Instance of `context_stack_base.ContextStack`, the context
-      stack to use for the bodies of the intrinsics.
-
-  Returns:
-    Instance of `computation_building_blocks.ComputationBuildingBlock` with all
-    the intrinsics from `intrinsic_bodies.py` inlined with their bodies, along
-    with a Boolean indicating whether there was any inlining in fact done.
-
-  Raises:
-    TypeError: If the types don't match.
-  """
-  py_typecheck.check_type(comp,
-                          computation_building_blocks.ComputationBuildingBlock)
-  py_typecheck.check_type(context_stack, context_stack_base.ContextStack)
-  bodies = intrinsic_bodies.get_intrinsic_bodies(context_stack)
-  transformed = False
-  for uri, body in six.iteritems(bodies):
-    comp, uri_found = replace_intrinsic_with_callable(comp, uri, body,
-                                                      context_stack)
-    transformed = transformed or uri_found
-  return comp, transformed
 
 
 def get_map_of_unbound_references(comp):
